hockey_api_scraper/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==22.0.0

SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1

requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
tenacity==8.5.0
